+++ src/libs/websockets/coincheck.py
@@ -1,35 +1,34 @@
-import json
-from websocket import create_connection
+import socketio
 
 from src.libs.websockets.websocket_client_base import WebsocketClientBase
-from src.constants.wsconst import WEBSOCKET_API_ENDPOINT_COINCHECK
-
-import src.utils.datetime as dt
+from src.constants.wsconst import SOCKETIO_URL
 
 
 class WebsocketClientCoincheck(WebsocketClientBase):
     def __init__(self, exchange_id, symbol):
         self.exchange_id = exchange_id
         self.symbol = symbol
-        self.ws = create_connection(WEBSOCKET_API_ENDPOINT_COINCHECK)
+
+        self.sio = socketio.Client(SOCKETIO_URL)
 
         symbols = symbol.split("/")
-        self.pair = "{}_{}".format(str.lower(symbols[0]),
+        self.PAIR = "{}_{}".format(str.lower(symbols[0]),
                                    str.lower(symbols[1]))
-        self.channel = "{}-orderbook".format(self.pair)
-        self.ws.send(json.dumps({
-            "type": "subscribe",
-            "channel": self.channel
-        }))
+        self.CHANNEL = "{}-orderbook".format(self.PAIR)
+
+    def connect(self):
+        self.sio.on('connect', self.on_connect)
+        self.sio.on('orderbook', self.on_orderbook)
+
+        self.sio.connect(SOCKETIO_URL,
+                         transports=['polling'],
+                         socketio_path='socket.io')
+
+    def on_orderbook(self, data):
+        print(data)
+
+    def on_connect(self):
+        self.sio.emit('subscribe', self.CHANNEL)
 
     def fetch_ticks(self):
-        while True:
-            timestamp = dt.now_timestamp_ms()
-            data = self.ws.recv()
-            data = json.loads(data)
-            ticks = {
-                "timestamp": timestamp,
-                "bids": data[1]["bids"],
-                "asks": data[1]["asks"]
-            }
-            print(ticks)
+        self.connect()
