+++ src/libs/websockets/coincheck_websocket.py
@@ -1,0 +1,35 @@
+import json
+from websocket import create_connection
+
+from src.libs.websockets.websocket_client_base import WebsocketClientBase
+from src.constants.wsconst import WEBSOCKET_API_ENDPOINT_COINCHECK
+
+import src.utils.datetime as dt
+
+
+class WebsocketClientCoincheck(WebsocketClientBase):
+    def __init__(self, exchange_id, symbol):
+        self.exchange_id = exchange_id
+        self.symbol = symbol
+        self.ws = create_connection(WEBSOCKET_API_ENDPOINT_COINCHECK)
+
+        symbols = symbol.split("/")
+        self.pair = "{}_{}".format(str.lower(symbols[0]),
+                                   str.lower(symbols[1]))
+        self.channel = "{}-orderbook".format(self.pair)
+        self.ws.send(json.dumps({
+            "type": "subscribe",
+            "channel": self.channel
+        }))
+
+    def fetch_ticks(self):
+        while True:
+            timestamp = dt.now_timestamp_ms()
+            data = self.ws.recv()
+            data = json.loads(data)
+            ticks = {
+                "timestamp": timestamp,
+                "bids": data[1]["bids"],
+                "asks": data[1]["asks"]
+            }
+            print(ticks)
